FdtLib: **TEST COMMIT** Error building EmbeddedPkg with CLANGPDB (only?)

+++ EmbeddedPkg/Library/FdtLib/fdt_strtoul.c
@@ -30,7 +30,7 @@ strtoul (
   }
 
   if (RETURN_ERROR (Status)) {
-    return MAX_UINTN;
+    return (unsigned long) MAX_UINTN;
   }
 
   return ReturnValue;
